CIT-32 création de l'entité et dtos

+++ src/main/java/com/wora/citronix/dtos/champ/ChampEmbeddedDTO.java
@@ -3,9 +3,10 @@ package com.wora.citronix.dtos.champ;
 import lombok.Data;
 import java.util.UUID;
 
+
 @Data
 public class ChampEmbeddedDTO {
     private UUID id;
-    private double superficie;
+    private Long superficie;
 }
