+++ src/main/java/com/wora/citronix/dtos/champ/ChampUpdateDTO.java
@@ -10,8 +10,9 @@ public class ChampUpdateDTO {
     private UUID id;
 
     @Positive(message = "La superficie doit être positive.")
-    private double superficie;
+    private Long superficie;
 
     @NotNull(message = "L'identifiant de la ferme est obligatoire.")
     private UUID fermeId;
+
 }
